Add files via upload
THE BOOK MANAGEMENT
Enables one to add new books with titles, authors and year of publication. 

BORROWING
Enables one to search for books using their titles,author and year of publication and allows for borrowing if the book is available. 
If the book has been borrowed already, it will direct the user to borrow another book. 
If the book is not available in the library, it gives feedback to the user to select another book. 
RETURNING BOOK
Allows the user to return a borrowed book. 
If the book had not been borrowed, the program informs the user that the book is not yet borrowed.